refactor: extract review orchestrator from webhook_handler (issue #61)
