+++ tests/github/test_reverify_threads.py
@@ -78,7 +78,7 @@ def _make_pr_context(awaiting_threads: list[DiscussionThread] | None = None) -> 
 @pytest.mark.asyncio
 async def test_reverify_fp_verdict_triggers_reply_and_resolve():
     """fp verdict → reply to thread + resolve."""
-    from baloo.github.webhook_handler import _reverify_awaiting_threads
+    from baloo.review.orchestrator import _reverify_awaiting_threads
 
     thread = _make_awaiting_thread(root_comment_id=42, node_id="PRT_x")
     pr_context = _make_pr_context(awaiting_threads=[thread])
@@ -104,7 +104,7 @@ async def test_reverify_fp_verdict_triggers_reply_and_resolve():
     mock_api.reply_to_review_comment = AsyncMock(return_value=True)
     mock_api.resolve_review_thread = AsyncMock(return_value=True)
 
-    with patch("baloo.github.webhook_handler.FPVerifier") as mock_verifier_cls:
+    with patch("baloo.review.orchestrator.FPVerifier") as mock_verifier_cls:
         mock_instance = AsyncMock()
         mock_instance.verify = AsyncMock(return_value=fp_result)
         mock_verifier_cls.return_value = mock_instance
@@ -128,7 +128,7 @@ async def test_reverify_fp_verdict_triggers_reply_and_resolve():
 @pytest.mark.asyncio
 async def test_reverify_real_verdict_no_action():
     """real verdict → thread untouched."""
-    from baloo.github.webhook_handler import _reverify_awaiting_threads
+    from baloo.review.orchestrator import _reverify_awaiting_threads
 
     thread = _make_awaiting_thread()
     pr_context = _make_pr_context(awaiting_threads=[thread])
@@ -144,7 +144,7 @@ async def test_reverify_real_verdict_no_action():
 
     mock_api = AsyncMock()
 
-    with patch("baloo.github.webhook_handler.FPVerifier") as mock_verifier_cls:
+    with patch("baloo.review.orchestrator.FPVerifier") as mock_verifier_cls:
         mock_instance = AsyncMock()
         mock_instance.verify = AsyncMock(return_value=fp_result)
         mock_verifier_cls.return_value = mock_instance
@@ -163,7 +163,7 @@ async def test_reverify_real_verdict_no_action():
 @pytest.mark.asyncio
 async def test_reverify_skips_threads_without_node_id():
     """Threads missing node_id are excluded."""
-    from baloo.github.webhook_handler import _reverify_awaiting_threads
+    from baloo.review.orchestrator import _reverify_awaiting_threads
 
     thread = _make_awaiting_thread(node_id=None)
     thread.node_id = None
@@ -171,7 +171,7 @@ async def test_reverify_skips_threads_without_node_id():
 
     mock_api = AsyncMock()
 
-    with patch("baloo.github.webhook_handler.FPVerifier") as mock_verifier_cls:
+    with patch("baloo.review.orchestrator.FPVerifier") as mock_verifier_cls:
         mock_instance = AsyncMock()
         mock_instance.verify = AsyncMock(return_value=FPVerificationResult())
         mock_verifier_cls.return_value = mock_instance
@@ -190,12 +190,12 @@ async def test_reverify_skips_threads_without_node_id():
 @pytest.mark.asyncio
 async def test_reverify_empty_list_no_verifier_call():
     """Empty awaiting_threads returns 0 immediately."""
-    from baloo.github.webhook_handler import _reverify_awaiting_threads
+    from baloo.review.orchestrator import _reverify_awaiting_threads
 
     pr_context = _make_pr_context()
     mock_api = AsyncMock()
 
-    with patch("baloo.github.webhook_handler.FPVerifier") as mock_verifier_cls:
+    with patch("baloo.review.orchestrator.FPVerifier") as mock_verifier_cls:
         resolved_count = await _reverify_awaiting_threads(
             awaiting_threads=[],
             pr_context=pr_context,
@@ -212,7 +212,7 @@ async def test_awaiting_count_excludes_resolved_threads():
     # This test verifies the integration: that auto_resolved_count returned from
     # _reverify_awaiting_threads correctly reduces the awaiting_threads count
     # fed to the decision engine, preventing spurious request_changes=True.
-    from baloo.github.webhook_handler import _reverify_awaiting_threads
+    from baloo.review.orchestrator import _reverify_awaiting_threads
 
     thread = _make_awaiting_thread(root_comment_id=99, node_id="PRT_awaiting")
     pr_context = _make_pr_context(awaiting_threads=[thread])
@@ -241,7 +241,7 @@ async def test_awaiting_count_excludes_resolved_threads():
     mock_api.reply_to_review_comment = AsyncMock(return_value=True)
     mock_api.resolve_review_thread = AsyncMock(return_value=True)
 
-    with patch("baloo.github.webhook_handler.FPVerifier") as mock_verifier_cls:
+    with patch("baloo.review.orchestrator.FPVerifier") as mock_verifier_cls:
         mock_instance = AsyncMock()
         mock_instance.verify = AsyncMock(return_value=fp_result)
         mock_verifier_cls.return_value = mock_instance
